[FE Feat]SessionChecking 추가
